% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{iris}
\alias{iris}
\title{Iris dataset}
\format{
An object of class \code{data.frame} with 150 rows and 5 columns.
}
\usage{
iris
}
\description{
Iris dataset
}
\keyword{datasets}
