% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_figure.R
\name{mod_download_figure_server}
\alias{mod_download_figure_server}
\title{download_figure Server Functions}
\usage{
mod_download_figure_server(
  id,
  filename,
  figure,
  width = 8,
  height = 6,
  label = "Download Plot"
)
}
\arguments{
\item{id}{String to identify namespace as an internal parameter for {shiny}.}

\item{filename}{The name of file that plot should be saved as}

\item{figure}{The plot to be downloaded. Should be wrapped in a reactive
function}

\item{width}{Width (inches) the plot should be saved as.
Default is 8 in and must be a value between 2 and 30.}

\item{height}{Height (inches) the plot should be saved as.
Default is 6 in and must be a value between 2 and 30.}

\item{label}{Character string denoting the label for the download button.
Default is "Download plot"}
}
\value{
{shiny} server function for download pop-up modal
}
\description{
The server function of the {shiny} module that works with
\code{\link{mod_download_figure_ui}()} to automatically download figures from
within a {shiny} application. The server function handles the UI for the
pop-up and code for downloading the plot as PDF, PNG, or SVG.  This works for
both ggplot2 graphics and base R plots. Base R plots must be save in an 
object with \code{\link{recordPlot}()}.
}
