% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_figure.R
\name{mod_download_figure_ui}
\alias{mod_download_figure_ui}
\title{download_figure UI Function}
\usage{
mod_download_figure_ui(id, label = "Download Plot")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{label}{Label for download button. Default is "Download Plot".}
}
\value{
UI elements for download button.
}
\description{
The UI function of the {shiny} module that works with 
mod_download_figure_server() to automatically download figures from within a 
{shiny} application. The UI function provides a "Download" button that 
results in a pop-up where users can enter plot height and width and options 
to download a plot as PDF, PNG, or SVG. This work for both {ggplot2} graphics 
and base R plots.
}
