% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_figure.R
\name{mod_download_figure_ui}
\alias{mod_download_figure_ui}
\title{download_figure UI Function}
\usage{
mod_download_figure_ui(id)
}
\arguments{
\item{id}{String to identify namespace as an internal parameter for {shiny}.}
}
\value{
UI elements for download button.
}
\description{
The UI function of the {shiny} module that works with 
\code{\link{mod_download_figure_server}()} to automatically download figures 
from within a {shiny} application. The UI function provides a "Download" 
button that results in a pop-up where users can enter plot height and width 
and options to download a plot as PDF, PNG, or SVG. This work for both 
{ggplot2} graphics and base R plots. Base R plots must be saved in an object 
with \code{\link{recordPlot}()}.
}
\seealso{
\code{\link{mod_download_figure_server}()} for server component and 
\code{\link{recordPlot}()} for additional documentation.
}
